dom/workers/test/test_notification.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=916893
-->
<head>
  <title>Bug 916893</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/dom/notification/test/mochitest/NotificationTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=916893">Bug 916893</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
<script type="text/javascript">
  SimpleTest.requestFlakyTimeout("Mock alert service dispatches show and click events.");

  async function runTest() {
    var w = new Worker("notification_worker.js");
    w.onmessage = function(e) {
      if (e.data.type === 'finish') {
        SimpleTest.finish();
      } else if (e.data.type === 'ok') {
        ok(e.data.test, e.data.message);
      } else if (e.data.type === 'is') {
        is(e.data.test1, e.data.test2, e.data.message);
      }
    }

    SimpleTest.waitForExplicitFinish();
    await NotificationTest.allowNotifications();
    w.postMessage('start')
  }

  SimpleTest.waitForExplicitFinish();
  runTest();
</script>
</body>
</html>
